+++ octasine/src/audio/parameters/operator_panning.rs
@@ -1,4 +1,4 @@
-use std::f64::consts::FRAC_PI_2;
+use std::f32::consts::FRAC_PI_2;
 
 use crate::audio::common::InterpolationDuration;
 use crate::common::SampleRate;
@@ -9,12 +9,12 @@ use super::common::{AudioParameter, Interpolator};
 #[derive(Debug, Clone)]
 pub struct OperatorPanningAudioParameter {
     value: Interpolator,
-    pub left_and_right: [f64; 2],
+    pub left_and_right: [f32; 2],
     pub lfo_active: bool,
 }
 
 impl OperatorPanningAudioParameter {
-    pub fn calculate_left_and_right(panning: f64) -> [f64; 2] {
+    pub fn calculate_left_and_right(panning: f32) -> [f32; 2] {
         let pan_phase = panning * FRAC_PI_2;
 
         [pan_phase.cos(), pan_phase.sin()]
@@ -43,7 +43,7 @@ impl AudioParameter for OperatorPanningAudioParameter {
     fn get_value(&self) -> <Self::ParameterValue as ParameterValue>::Value {
         self.value.get_value()
     }
-    fn set_from_patch(&mut self, value: f64) {
+    fn set_from_patch(&mut self, value: f32) {
         self.value
             .set_value(Self::ParameterValue::new_from_patch(value).get())
     }
@@ -55,7 +55,7 @@ impl AudioParameter for OperatorPanningAudioParameter {
             let patch_value = Self::ParameterValue::new_from_audio(self.get_value()).to_patch();
 
             let new_panning = Self::ParameterValue::new_from_patch(
-                (patch_value + lfo_addition as f64).min(1.0).max(0.0),
+                (patch_value + lfo_addition as f32).min(1.0).max(0.0),
             )
             .get();
